third_party/blink/web_tests/platform/fuchsia/navigator_language/navigator_language-expected.txt

This is a testharness.js-based test.
[FAIL] Test that navigator.languages reflects the accept languages value.
  assert_equals: expected 1 but got 2
Harness: the test ran to completion.
